+++ extension.js
@@ -26,25 +26,39 @@ function activate(context) {
   );
 
   context.subscriptions.push(
-    vscode.commands.registerTextEditorCommand(
+    vscode.commands.registerCommand(
       'closeTag.closeHTMLTagInPlace',
-      (textEditor, edit, args) => {
-        const newSelections = textEditor.selections;
+      async (args) => {
+        const { activeTextEditor: textEditor } = vscode.window;
 
-        closeSelections(textEditor, edit, args);
+        if (textEditor) {
+          const prevSelections = textEditor.selections;
 
-        textEditor.selections = newSelections;
+          await textEditor.edit(edit => {
+            closeSelections(textEditor, edit, args);
+          });
+
+          prevSelections.forEach((selection, index) => {
+            // If a range of text was selected previously, then just keep it
+            // since the new replacement text will be selected already.
+            // Otherwise, restore the previous cursor position.
+
+            if (selection.isEmpty) {
+              textEditor.selections[index] = selection;
+            }
+          });
+        }
       }
     )
   );
 }
 
-function closeSelections(textEditor, edit, overrideOptions = {}) {
+function closeSelections(textEditor, edit, overrideOptions) {
   const ignoreTags = reduceMap(
     Object.assign(
       {},
       normalizeIgnoreTags(vscode.workspace.getConfiguration('closeTag').get('ignoreTags')),
-      normalizeIgnoreTags(overrideOptions.ignoreTags)
+      normalizeIgnoreTags((overrideOptions || {}).ignoreTags)
     ),
     (result, value, name) => {
       value && result.push(name);